backend/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
sqlalchemy==2.0.27
postgresql+psycopg://neondb_owner:npg_Fl0Darhg5IRV@ep-shy-heart-adkufq2e-pooler.c-2.us-east-1.aws.neon.tech/neondb?sslmode=require

pgvector==0.2.5
pydantic==1.10.14
python-dotenv==1.0.1
numpy==1.26.4
python-dateutil==2.8.2
pytest==7.4.4
